# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fixly API
  version: 1.0.0
  description: |
    Public + admin HTTP API for Fixly. All responses follow the project's
    envelope contract: `{ "data": ... }` on success and
    `{ "error": { "code": string, "message": string } }` on failure. The
    socket.io surface (`booking:status`, `location:update`) is documented
    under the `Realtime` tag.
servers:
  - url: /api/v1
    description: Same-origin (proxied via Vite in dev, ingress in prod)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: phone must be a valid mobile number }
...
    Service:
      type: object
      properties:
        id: { type: string, example: elec }
        nameAr: { type: string, example: كهرباء }
        nameEn: { type: string, example: Electricity }
        descriptionAr: { type: string, nullable: true }
        priceJod: { type: string, example: '50.00' }
        durationMin: { type: integer, example: 45 }
        isActive: { type: boolean, example: true }
        category: { type: string, nullable: true, example: HOME_REPAIR }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [AWAITING_PAYMENT, PENDING, CONFIRMED, EN_ROUTE, ARRIVED, IN_PROGRESS, COMPLETED, CANCELLED, DISPUTED, NO_SHOW]
        scheduledAt: { type: string, format: date-time, nullable: true }
        totalJod: { type: string, example: '50.00' }
        service: { $ref: '#/components/schemas/Service' }
...
    BookingQuote:
      type: object
      description: >-
        Video pre-check (fixed_scope convenience) OR itemized quote-first
        (mandatory for quote_first services, §17.5) — `lines` is empty for the
        former, populated for the latter; labourFils/materialsFils are summed
        from `lines` for the itemized path.
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        serviceId: { type: string }
        status: { type: string, enum: [PENDING, QUOTED, ACCEPTED, DECLINED, EXPIRED] }
        videoUrl: { type: string, format: uri, nullable: true }
        siteMediaUrls: { type: array, items: { type: string, format: uri } }
        dimensionsNote: { type: string, nullable: true }
        requestedTier: { type: string, nullable: true, enum: [ECONOMY, STANDARD, PREMIUM] }
        description: { type: string, nullable: true }
        quotedJod: { type: string, nullable: true, example: '105.000' }
        labourFils: { type: integer, nullable: true, example: 45000 }
        materialsFils: { type: integer, nullable: true, example: 60000 }
        opsReviewedById: { type: string, format: uuid, nullable: true }
        opsReviewedAt: { type: string, format: date-time, nullable: true }
        lines: { type: array, items: { $ref: '#/components/schemas/QuoteLine' } }
        addressLine: { type: string, nullable: true }
        addressLat: { type: number, nullable: true }
        addressLng: { type: number, nullable: true }
        bookingId: { type: string, format: uuid, nullable: true }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    QuoteLine:
      type: object
      description: One itemized labour/material/prep/fee line on a quote-first BookingQuote (§17.5.3).
      properties:
        id: { type: string, format: uuid }
        quoteId: { type: string, format: uuid }
        kind: { type: string, enum: [LABOUR, MATERIAL, PREP, FEE] }
        materialId: { type: string, format: uuid, nullable: true, description: References MaterialCatalog; null = off-catalogue (routes to ops review). }
        description: { type: string }
        qty: { type: string, example: '2.00' }
        unit: { type: string, nullable: true, example: bucket }
        unitPriceFils: { type: integer, example: 12000 }
        totalFils: { type: integer, example: 24000 }
        source: { type: string, enum: [TECHNICIAN_PROCURED, CUSTOMER_SUPPLIED, PLATFORM_ARRANGED] }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /quotes:
    post:
      tags: [Quotes]
      summary: Request a firm quote — video pre-check (fixed_scope) or quote-first assessment (mandatory for quote_first services)
      description: >-
        Exactly one of videoUrl (fixed_scope) or siteMediaUrls (quote_first) is
        required, decided by the target service's pricingModel — enforced
        server-side regardless of which is sent.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceId]
              properties:
                serviceId: { type: string, format: uuid }
                videoUrl: { type: string, format: uri, maxLength: 500, description: HTTPS URL only; required for a fixed_scope service. }
                siteMediaUrls: { type: array, maxItems: 20, items: { type: string, format: uri }, description: HTTPS URLs; required for a quote_first service. }
                dimensionsNote: { type: string, maxLength: 1000 }
                requestedTier: { type: string, enum: [ECONOMY, STANDARD, PREMIUM] }
                description: { type: string, maxLength: 1000 }
                addressLine: { type: string, maxLength: 500 }
                addressLat: { type: number, minimum: -90, maximum: 90 }
                addressLng: { type: number, minimum: -180, maximum: 180 }
      responses:
        '201':
          description: Quote created (status PENDING).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/BookingQuote' } } }
        '404': { description: Service not found. }
        '409': { description: Service is not active. }
        '422': { description: Wrong intake shape for this service's pricing archetype. }
    get:
      tags: [Quotes]
      summary: My video pre-check quotes
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Quotes.
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/BookingQuote' } } } }
  /quotes/{id}:
    get:
      tags: [Quotes]
      summary: A quote I own
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Quote.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/BookingQuote' } } }
        '404': { $ref: '#/components/schemas/Error' }
  /quotes/{id}/accept:
    post:
      tags: [Quotes]
      summary: Accept a firm quote — creates a booking at exactly quotedJod
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '201':
          description: Booking created at the quoted price.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Booking' } } }
        '403': { $ref: '#/components/schemas/Error' }
        '409': { description: Quote is not QUOTED, already accepted, or its expiresAt has passed. }

  # ---- Bill of Materials (§17.5.9) — technician-managed, customer ack/decline ----
...
  /admin/quotes:
    get:
      tags: [Admin]
      summary: Video pre-check quote queue
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [PENDING, QUOTED, ACCEPTED, DECLINED, EXPIRED] } }
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0, minimum: 0 } }
      responses: { '200': { description: Paginated quotes. } }
  /admin/quotes/{id}/quote:
    post:
      tags: [Admin]
      summary: Set a firm price on a pending quote (OPS)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quotedJod]
              properties:
                quotedJod: { type: string, description: 'Positive amount, max 3 decimal places (fils).' }
      responses:
        '200':
          description: Quoted (status QUOTED); customer notified.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/BookingQuote' } } }
        '409': { description: Quote is not PENDING. }

  # ---- Admin: itemized quote-line drafting (OPS) — §17.5.3/§17.5.12 ----
...